# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ERP Management API",
    "description": "ERP后台接口文档，当前包含认证系统与客户管理模块。",
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/dashboard/boss": {
      "get": {
        "tags": ["老板驾驶舱"],
        "summary": "老板驾驶舱经营数据",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "success" } }
      }
    },
...
  },
...
}